refactor(app): move SchulCloudApp to module root
