+++ lib/app/app.dart
@@ -23,6 +23,7 @@ export 'error_reporting.dart';
 export 'hive.dart';
 export 'logger.dart';
 export 'routing.dart' show FancyRoute, FancyRouteBuilder;
+export 'schulcloud_app.dart' show SchulCloudApp, SignedInScreen;
 export 'services/api_network.dart';
 export 'services/deep_linking.dart';
 export 'services/network.dart';
@@ -37,5 +38,4 @@ export 'widgets/app_bar.dart';
 export 'widgets/fade_in.dart';
 export 'widgets/form.dart';
 export 'widgets/scaffold.dart';
-export 'widgets/schulcloud_app.dart' show SchulCloudApp, SignedInScreen;
 export 'widgets/text.dart';
