+++ lib/app/routing.dart
@@ -10,10 +10,10 @@ import 'package:schulcloud/settings/settings.dart';
 import 'package:schulcloud/sign_in/sign_in.dart';
 import 'package:swipeable_page_route/swipeable_page_route.dart';
 
+import 'schulcloud_app.dart';
 import 'top_level_route/page_route.dart';
 import 'utils.dart';
 import 'widgets/not_found_screen.dart';
-import 'widgets/schulcloud_app.dart';
 
 final hostRegExp = RegExp('(?:www\.)?${RegExp.escape(services.config.host)}');
